Development API certified

+++ BikeRental.Core/Bike.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
         public Status Status { get; set; }
-        public ushort Price { get; set; }
+        public float Price { get; set; }
 
         public BikeType BikeType { get; set; }
         public int BikeTypeId { get; set; }
